requirements.txt
fastapi[standard]==0.112.1
pydantic==2.8.0
torch
flashrank
langchain-openai
langchain-community
faiss-cpu
redisvl
presidio-analyzer # For Sensitive data masking
presidio-anonymizer 
spacy
nemoguardrails
promptflow-tracing
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation-langchain
ragas
sentence-transformers
rank_bm25